editor/libeditor/tests/test_bug1649005.html

<!DOCTYPE>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1649005
-->
<head>
  <meta charset="UTF-8" />
  <title>Test for bug 1649005</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" href="/tests/SimpleTest/test.css">
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script>
  /** Test for bug 1649005, bug 1779343 **/
  window.addEventListener("DOMContentLoaded", (event) => {
    SimpleTest.waitForExplicitFinish();
    SimpleTest.waitForFocus(function() {
      document.body.textContent = ""; // It would be \n\n otherwise...
      synthesizeMouseAtCenter(document.body, {});

      var editor = getEditor();
      is(document.body.textContent, "", "Initial body check");
      editor.rewrap(false);
      is(document.body.textContent, "", "Initial body check after rewrap");

      document.body.innerHTML = "&gt;abc<br/>&gt;def<br/>&gt;ghi";
      editor.rewrap(true);
      is(document.body.textContent, "> abc def ghi", "Rewrapped");

      document.body.innerHTML = "&gt; ";
      editor.rewrap(true);
      is(document.body.textContent, "> ", "Rewrapped half-empty string");

      SimpleTest.finish();
    });
  });

  function getEditor() {
    var Ci = SpecialPowers.Ci;
    var editingSession = SpecialPowers.wrap(window).docShell.editingSession;
    var editor = editingSession.getEditorForWindow(window);
    editor.QueryInterface(Ci.nsIHTMLEditor);
    editor.QueryInterface(Ci.nsIEditorMailSupport);
    editor.flags |= SpecialPowers.Ci.nsIEditor.eEditorPlaintextMask;
    return editor;
  }
  </script>
</head>
<body contenteditable></body>
</html>
